refactor: 파일, 폴더구조 리팩토링
