+++ src/service/product.type.ts
@@ -1,4 +1,4 @@
-import { BaseData } from "./common";
+import { BaseData } from "../types/common";
 
 export type Tags = string[];
